Разобрал классы по отдельным модулям, и отрисовал коммиты по ключу
