+++ src/fish/card.js
@@ -1,22 +1,16 @@
-import {randomData} from './util.js';
-import {ID, POSTERS, TITLE_FILMS, DESCRIPTION_FILMS, AGE_RATING, GENRE, ALTERNATIVE_TITLE, RUNTIME, RELEASE_COUNTRY, WATCHING_DATE} from './data.js';
+import {getRandomData} from './util.js';
+import {ID, POSTERS, TITLES_FILMS, DESCRIPTION_FILMS, AGE_RATINGS, GENRES, ALTERNATIVE_TITLES, RUNTIMES, RELEASE_COUNTRYS, WATCHING_DATES} from './data.js';
 
 
-const getFilmCards = () => ({
-  id: randomData(ID),
-  comments: [
-    {id: 1},
-    {id: 2},
-    {id: 3},
-    {id: 4},
-    {id: 5}
-  ],
+const getFilmCard = () => ({
+  id: getRandomData(ID),
+  comments: [1, 2, 3, 4, 5],
   filmInfo: {
-    title: randomData(TITLE_FILMS), //заголовок
-    alternativeTitle: randomData(ALTERNATIVE_TITLE), //Альтернативный заголовок
+    title: getRandomData(TITLES_FILMS), //заголовок
+    alternativeTitle: getRandomData(ALTERNATIVE_TITLES), //Альтернативный заголовок
     totalRating: '99.4', //Общая оценка
-    poster: `./images/posters/${randomData(POSTERS)}`, //плакат
-    ageRating: randomData(AGE_RATING), //возрастной рейтинг
+    poster: `./images/posters/${getRandomData(POSTERS)}`, //плакат
+    ageRating: getRandomData(AGE_RATINGS), //возрастной рейтинг
     director: 'Конено, Я!',
     writers: [ //Сценаристы
       'Тоже, Я!'
@@ -25,21 +19,21 @@ const getFilmCards = () => ({
       'Самый крутой актёр это кто? Никогда не догадаетесь... Тадам, это - Я'
     ],
     release: { //выпущен
-      date: randomData(WATCHING_DATE),
-      releaseCountry: randomData(RELEASE_COUNTRY), //Страна выпуска
+      date: getRandomData(WATCHING_DATES),
+      releaseCountry: getRandomData(RELEASE_COUNTRYS), //Страна выпуска
     },
-    runtime: randomData(RUNTIME), //продолжительность фильма
+    runtime: getRandomData(RUNTIMES), //продолжительность фильма
     genre: [ //жанр
-      randomData(GENRE)
+      getRandomData(GENRES)
     ],
-    description: randomData(DESCRIPTION_FILMS), //описание
+    description: getRandomData(DESCRIPTION_FILMS), //описание
   },
   userDetails: {
     watchlist: false, //список наблюдения
     alreadyWatched: true, //уже Смотрел
-    watchingDate: randomData(WATCHING_DATE), //Дата просмотра
+    watchingDate: getRandomData(WATCHING_DATES), //Дата просмотра
     favorite: false, //любимый фильм
   }
 });
 
-export {getFilmCards};
+export {getFilmCard};
