+++ src/fish/comment.js
@@ -1,20 +1,8 @@
-import {randomData} from '../fish/util.js';
-import {ID, AUTHOR, COMMENT, WATCHING_DATE} from '../fish/data.js';
-
 const localComment = {
   comment: 'a film that changed my life, a true masterpiece, post-credit scene was just amazing omg.',
   emotion: 'smile'
 };
 
-const EMOTION = ['smile', 'sleeping', 'puke', 'angry'];
+const EMOTIONS = ['smile', 'sleeping', 'puke', 'angry'];
 
-
-const getArrayComments = () =>({
-  id: randomData(ID),
-  author: randomData(AUTHOR),
-  comment: randomData(COMMENT),
-  date: randomData(WATCHING_DATE),
-  emotion: randomData(EMOTION)
-});
-
-export {localComment, EMOTION, getArrayComments};
+export {localComment, EMOTIONS};
