+++ src/fish/util.js
@@ -7,7 +7,7 @@ const getRandomInteger = (a = 0, b = 1) => {
   return Math.floor(lower + Math.random() * (upper - lower + 1));
 };
 
-const randomData = (data) => {
+const getRandomData = (data) => {
   const randomIndex = getRandomInteger(0, data.length - 1);
   return data[randomIndex];
 };
@@ -17,4 +17,4 @@ const isCommentFilm = (comment) => Object.values(comment).some(Boolean);
 const humanizeReleaseDate = (dueDate) => dayjs(dueDate).format('D MMMM YYYY');
 const humanizeDate = (date) => dayjs(date).format('YYYY');
 
-export {getRandomInteger, humanizeReleaseDate, randomData, humanizeDate, isCommentFilm};
+export {getRandomInteger, humanizeReleaseDate, getRandomData, humanizeDate, isCommentFilm};
