+++ src/model/film-cards-model.js
@@ -1,13 +1,14 @@
-import {getFilmCards} from '../fish/card.js';
-import {getArrayComments} from '../fish/comment.js';
+import {getFilmCard} from '../fish/card.js';
+import {commentsArray} from '../fish/data.js';
+import {filterComments} from '../view/film-details-comment-view.js';
 export default class FilmCardModel {
-  filmCard = Array.from({length: 5}, getFilmCards);
+  filmCard = Array.from({length: 5}, getFilmCard);
   getFilmCard = () => this.filmCard;
 
-  filmCardPopup = Array.from({length: 1}, getFilmCards);
+  filmCardPopup = Array.from({length: 1}, getFilmCard);
   getFilmCardPopup = () => this.filmCardPopup;
 
-  filmDetailsComment = Array.from({length: 5}, getArrayComments);
+  filmDetailsComment = filterComments(commentsArray, getFilmCard);
   getComments = () => this.filmDetailsComment;
 }
