+++ src/view/films-list-container-view.js
@@ -1,0 +1,21 @@
+import {createElement} from '../render.js';
+
+const createMovieCardContainerViewTemplate = () => '<div class="films-list__container"></div>';
+
+export default class MovieCardContainerView {
+  getTemplate() {
+    return createMovieCardContainerViewTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
