+++ src/view/films-list-show-more-button-view.js
