+++ src/view/films-list-view.js
@@ -1,0 +1,24 @@
+import {createElement} from '../render.js';
+
+const createMovieCardListViewTemplate = () => (
+  `<section class="films-list">
+    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+    </section>`
+);
+export default class MovieCardListView {
+  getTemplate() {
+    return createMovieCardListViewTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
