+++ src/view/filters-view.js
@@ -1,32 +1,5 @@
 import {createElement} from '../render.js';
 
-const createFiltersNavigationViewTemplate = () =>
-  (`<nav class="main-navigation">
-  <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-  <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-  <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-  <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-</nav>`
-  );
-class FiltersNavigationView {
-  getTemplate() {
-    return createFiltersNavigationViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
 const createFiltersSortViewTemplate = () =>
   (`
 <ul class="sort">
@@ -35,7 +8,7 @@ const createFiltersSortViewTemplate = () =>
     <li><a href="#" class="sort__button">Sort by rating</a></li>
   </ul>`
   );
-class FiltersSortView {
+export default class FiltersSortView {
   getTemplate() {
     return createFiltersSortViewTemplate();
   }
@@ -52,5 +25,3 @@ class FiltersSortView {
     this.element = null;
   }
 }
-
-export {FiltersNavigationView, FiltersSortView};
