--- src/view/movie-card-view.js
@@ -1,261 +1,0 @@
-import {createElement} from '../render.js';
-import {humanizeDate} from '../fish/util.js';
-
-const createFilmsListViewTemplate = () => '<section class="films"></section>';
-class FilmsListView {
-  getTemplate() {
-    return createFilmsListViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
-const createMovieCardListViewTemplate = () => (
-  `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-    </section>`
-);
-class MovieCardListView {
-  getTemplate() {
-    return createMovieCardListViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
-const createMovieCardContainerViewTemplate = () => '<div class="films-list__container"></div>';
-class MovieCardContainerView {
-  getTemplate() {
-    return createMovieCardContainerViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-const createMovieCardViewTemplate = (film) => {
-  const {title, release, poster, ageRating, runtime, genre, description} = film.filmInfo;
-  const MAX_DESCRIPTION_LENGTH = 140;
-
-  const years = release.date !== null
-    ? humanizeDate(release.date)
-    : '';
-
-  const createDescription = (text) => {
-    if (text.length < MAX_DESCRIPTION_LENGTH) {
-      return text;
-    }
-
-    return `${text.slice(0, MAX_DESCRIPTION_LENGTH - 2)}...`;
-  };
-
-  return (
-    `<article class="film-card">
-<a class="film-card__link">
-  <h3 class="film-card__title">${title}</h3>
-  <p class="film-card__rating">${ageRating}</p>
-  <p class="film-card__info">
-    <span class="film-card__year">${years}</span>
-    <span class="film-card__duration">${runtime}</span>
-    <span class="film-card__genre">${genre}</span>
-  </p>
-  <img src="${poster}" alt="" class="film-card__poster">
-  <p class="film-card__description">${createDescription(description)}</p>
-  <span class="film-card__comments">30 comments</span>
-
-</a>
-<div class="film-card__controls">
-  <button class="film-card__controls-item film-card__controls-item--add-to-watchlist film-card__controls-item--active" type="button">Add to watchlist</button>
-  <button class="film-card__controls-item film-card__controls-item--mark-as-watched film-card__controls-item--active" type="button">Mark as watched</button>
-  <button class="film-card__controls-item film-card__controls-item--favorite film-card__controls-item--active" type="button">Mark as favorite</button>
-</div>
-</article>`);
-};
-class MovieCardView {
-  constructor (film){
-    this.film = film;
-  }
-
-  getTemplate() {
-    return createMovieCardViewTemplate(this.film);
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
-const createMovieCardExtraTopRatedViewTemplate = () =>(
-  `<section class="films-list films-list--extra">
-    <h2 class="films-list__title">Top rated</h2>
-
-    <div class="films-list__container">
-      <article class="film-card">
-        <a class="film-card__link">
-          <h3 class="film-card__title">The Man with the Golden Arm</h3>
-          <p class="film-card__rating">9.0</p>
-          <p class="film-card__info">
-            <span class="film-card__year">1955</span>
-            <span class="film-card__duration">1h 59m</span>
-            <span class="film-card__genre">Drama</span>
-          </p>
-          <img src="./images/posters/the-man-with-the-golden-arm.jpg" alt="" class="film-card__poster">
-          <p class="film-card__description">Frankie Machine (Frank Sinatra) is released from the federal Narcotic Farm in Lexington, Kentucky with a set of drums and a new outlook on…</p>
-          <span class="film-card__comments">18 comments</span>
-        </a>
-        <div class="film-card__controls">
-          <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
-          <button class="film-card__controls-item film-card__controls-item--mark-as-watched film-card__controls-item--active" type="button">Mark as watched</button>
-          <button class="film-card__controls-item film-card__controls-item--favorite" type="button">Mark as favorite</button>
-        </div>
-      </article>
-
-      <article class="film-card">
-        <a class="film-card__link">
-          <h3 class="film-card__title">The Great Flamarion</h3>
-          <p class="film-card__rating">8.9</p>
-          <p class="film-card__info">
-            <span class="film-card__year">1945</span>
-            <span class="film-card__duration">1h 18m</span>
-            <span class="film-card__genre">Mystery</span>
-          </p>
-          <img src="./images/posters/the-great-flamarion.jpg" alt="" class="film-card__poster">
-          <p class="film-card__description">The film opens following a murder at a cabaret in Mexico City in 1936, and then presents the events leading up to it in flashback. The Grea…</p>
-          <span class="film-card__comments">12 comments</span>
-        </a>
-        <div class="film-card__controls">
-          <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
-          <button class="film-card__controls-item film-card__controls-item--mark-as-watched" type="button">Mark as watched</button>
-          <button class="film-card__controls-item film-card__controls-item--favorite" type="button">Mark as favorite</button>
-        </div>
-      </article>
-    </div>
-  </section>`
-);
-class MovieCardExtraTopRatedView {
-  getTemplate() {
-    return createMovieCardExtraTopRatedViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
-const createMovieCardExtraMostCommentedViewTemplate = () =>(
-  `<section class="films-list films-list--extra">
-    <h2 class="films-list__title">Most commented</h2>
-
-    <div class="films-list__container">
-      <article class="film-card">
-        <a class="film-card__link">
-          <h3 class="film-card__title">Santa Claus Conquers the Martians</h3>
-          <p class="film-card__rating">2.3</p>
-          <p class="film-card__info">
-            <span class="film-card__year">1964</span>
-            <span class="film-card__duration">1h 21m</span>
-            <span class="film-card__genre">Comedy</span>
-          </p>
-          <img src="./images/posters/santa-claus-conquers-the-martians.jpg" alt="" class="film-card__poster">
-          <p class="film-card__description">The Martians Momar ("Mom Martian") and Kimar ("King Martian") are worried that their children Girmar ("Girl Martian") and Bomar ("Boy Marti…</p>
-          <span class="film-card__comments">465 comments</span>
-        </a>
-        <div class="film-card__controls">
-          <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
-          <button class="film-card__controls-item film-card__controls-item--mark-as-watched" type="button">Mark as watched</button>
-          <button class="film-card__controls-item film-card__controls-item--favorite film-card__controls-item--active" type="button">Mark as favorite</button>
-        </div>
-      </article>
-
-      <article class="film-card">
-        <a class="film-card__link">
-          <h3 class="film-card__title">Made for Each Other</h3>
-          <p class="film-card__rating">5.8</p>
-          <p class="film-card__info">
-            <span class="film-card__year">1939</span>
-            <span class="film-card__duration">1h 32m</span>
-            <span class="film-card__genre">Comedy</span>
-          </p>
-          <img src="./images/posters/made-for-each-other.png" alt="" class="film-card__poster">
-          <p class="film-card__description">John Mason (James Stewart) is a young, somewhat timid attorney in New York City. He has been doing his job well, and he has a chance of bei…</p>
-          <span class="film-card__comments">56 comments</span>
-        </a>
-        <div class="film-card__controls">
-          <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
-          <button class="film-card__controls-item film-card__controls-item--mark-as-watched" type="button">Mark as watched</button>
-          <button class="film-card__controls-item film-card__controls-item--favorite" type="button">Mark as favorite</button>
-        </div>
-      </article>
-    </div>
-  </section>
-</section>`
-);
-class MovieCardExtraMostCommentedView {
-  getTemplate() {
-    return createMovieCardExtraMostCommentedViewTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
-
-
-export{MovieCardListView, FilmsListView, MovieCardContainerView, MovieCardView, MovieCardExtraTopRatedView, MovieCardExtraMostCommentedView};
